Split carousel code into separate component
