+++ packages/web/src/pages/explore-page/components/desktop/constants.ts
@@ -1,0 +1,2 @@
+export const TILE_WIDTH = 532
+export const MOBILE_TILE_WIDTH = 343
